+++ packages/@expo/cli/src/utils/createFileTransform.ts
@@ -1,48 +1,9 @@
 import { IOSConfig } from '@expo/config-plugins';
-import Minipass from 'minipass';
 import path from 'path';
 import picomatch from 'picomatch';
 import { ReadEntry } from 'tar';
 
 const debug = require('debug')('expo:file-transform') as typeof console.log;
-
-function escapeXMLCharacters(original: string): string {
-  const noAmps = original.replace('&', '&amp;');
-  const noLt = noAmps.replace('<', '&lt;');
-  const noGt = noLt.replace('>', '&gt;');
-  const noApos = noGt.replace('"', '\\"');
-  return noApos.replace("'", "\\'");
-}
-
-class Transformer extends Minipass {
-  data = '';
-
-  constructor(private settings: { name: string; extension: string }) {
-    super();
-  }
-
-  write(data: string) {
-    this.data += data;
-    return true;
-  }
-
-  getNormalizedName(): string {
-    if (['.xml', '.plist'].includes(this.settings.extension)) {
-      return escapeXMLCharacters(this.settings.name);
-    }
-    return this.settings.name;
-  }
-
-  end() {
-    const name = this.getNormalizedName();
-    const replaced = this.data
-      .replace(/Hello App Display Name/g, name)
-      .replace(/HelloWorld/g, IOSConfig.XcodeUtils.sanitizedName(name))
-      .replace(/helloworld/g, IOSConfig.XcodeUtils.sanitizedName(name.toLowerCase()));
-    super.write(replaced);
-    return super.end();
-  }
-}
 
 export function createEntryResolver(name: string) {
   return (entry: ReadEntry) => {
@@ -62,37 +23,6 @@ export function createEntryResolver(name: string) {
       // See: https://github.com/npm/npm/issues/1862
       entry.path = entry.path.replace(/gitignore$/, '.gitignore');
     }
-  };
-}
-
-export function createFileTransform(name: string) {
-  return (entry: ReadEntry) => {
-    const extension = path.extname(entry.path);
-
-    // Binary files, don't process these (avoid decoding as utf8)
-    if (
-      ![
-        '.png',
-        '.jpg',
-        '.jpeg',
-        '.gif',
-        '.webp',
-        '.psd',
-        '.tiff',
-        '.svg',
-        '.jar',
-        '.keystore',
-        '.gz',
-
-        // Font files
-        '.otf',
-        '.ttf',
-      ].includes(extension) &&
-      name
-    ) {
-      return new Transformer({ name, extension });
-    }
-    return undefined;
   };
 }